// packages/hpl-payment-manager/generated/types/Payment.ts
/**
 * This code was GENERATED using the solita package.
 * Please DO NOT EDIT THIS FILE, instead rerun solita to update it or write a wrapper to add functionality.
 *
 * See: https://github.com/metaplex-foundation/solita
 */

import * as beet from '@metaplex-foundation/beet'
import { PaymentMethod, paymentMethodBeet } from './PaymentMethod'
import { PaymentKind, paymentKindBeet } from './PaymentKind'
export type Payment = {
  paymentMethod: PaymentMethod
  kind: PaymentKind
}

/**
 * @category userTypes
 * @category generated
 */
export const paymentBeet = new beet.FixableBeetArgsStruct<Payment>(
  [
    ['paymentMethod', paymentMethodBeet],
    ['kind', paymentKindBeet],
  ],
  'Payment'
)